+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Nothing yet!
+* 新增 graphQLAPI 的 headers 參數
 
 ## v1.2.4 - 2025-03-25